requirements.txt
# AI Director — Dependencies
# Core
pyyaml>=6.0.1
aiohttp>=3.9.0
requests>=2.31.0

# Discord
discord.py>=2.3.0

# ComfyUI Communication
websockets>=12.0

# Image handling
Pillow>=10.1.0

# LLM (Ollama is HTTP-based, no extra package needed)
# For Gemini
google-genai>=0.2.0
